Rename socket manager file
