+++ src/SocketManager.ts
